+++ src/Repository/Contract/ProjectRepositoryContract.php
@@ -2,4 +2,7 @@
 
 namespace App\Repository\Contract;
 
-interface ProjectRepositoryContract extends AbstractRepositoryContract {}
+interface ProjectRepositoryContract extends AbstractRepositoryContract
+{
+    public function findTeams(int $projectId, bool $active = true): ?array;
+}