+++ CHANGELOG.md
@@ -15,3 +15,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Create console commands: `app:project:new`, `app:project:list`, `app:project:view`, `app:project:assign-pm`
 - Create console commands: `app:user:list-tasks`
 - Create console commands: `app:task:list`, `app:task:new`, `app:task:list-expired`
+- Create console commands: `app:project:list-crossed`